Bass: Modify allowed context types at the beginning of add source
Disallowing SOUND_EFFECTS causes that even click to addSource not
generate touch sound, which is intended.

If context types are modified not at the beginning of addSource but
directly before the call to StateMachine and there is a need to sync
to broadcaster first, a sound effect is generated, which leads to
releasing ASE by buds when source is added.

This CL fixes that by moving modification of context types to the
beginning of addSource. Revert of that is done on each
notifySourceAddFailed so there is no risk that we left it changed
on fail.

Bug: 413301516
Flag: EXEMPT, day-to-day fix covered by unit tests
Test: atest BassClientServiceTest
Change-Id: I486b4c4ffe8a26a26af94a74327581349f4c1ffc