Block enable call if user hasn't started yet
Bluetooth must have an user to start for. On HSUM device, the root user
may not be the foreground user and instead, another user is selected to
start for.

When this happen, we need to know which user, and to initialize various
settings for this user specifically.

Thus, we absolutely cannot start without this user.

Instead of crashing randomly inside the bluetooth system server, make it
a clear rejection at method entry.

Bug: 432337346
Flag: EXEMPT bugfix
Test: m .
Change-Id: I48e6af77ef2bea7d3de010084c840e2a59e785bb