codec_manager: Fix possible crash when provider_info is std::nullopt
When platform is configured to not use Codec Extensibility, the
codec_provider_info_ is set to std::nullopt.

However, in some part of the code, we are checking
codec_provider_info_->isMulticodecSupported, without this checking it
codec_provider_info_ is valid.

This patch fixes that

Bug: 431668632
Flag: EXEMPT, day-to-day obvious fix.
Test: atest bluetooth_le_audio_codec_manager_test

Change-Id: Ic3fe33a46b7f966acedbd260be39d1d058691ca9