LeAudioService: Improve configuration change
When configuration of the LeAudio changes, LeAudioService is notified.
So far, whenever codec or sample frequence changed, the Audio Framework
was notified and because of this Audio Framework proceed with Audio
Device TearDown and SetUp.
However, at least for the LC3 codec type, whenever sample frequency
changes, there is no need for this kind of reset from the Audio
Framework point of view. This is because, Bluetooth already created a
session with a Sample Frequecny higier than needed and downsampling is
happening in the Audio HAL - and Audio HAL knows the configuration from
Bluetooth.

With this CL, Audio Framework is notified about configuration change
when codec changes and also when sample frequency changes for other
codecs than LC3.

Bug: 432126145
Flag: EXEMPT, day-to-day fix
Test: atest bluetooth_le_audio_client_test
Change-Id: I09d069416f8fd4fdfe34c4d5c2ca7b9af29bf939