Socket: less getAddress, more attributionSource
The Socket mAddress field is not used. There is no point in even trying
to get the local mac address since we are not using it. So deleting this
field entirely.
This is raising a second issue, that the BLUETOOTH_CONNECT permission
wasn't enforce properly were we thought it was enforced. So we provide
the attributionSource of the caller to validate it has permissions.
Finally, make the BluetoothDevice of the socket final, and optional as
there are no device when this is a server socket

Flag: EXEMPT refactor no-op
Bug: 385055956
Test: atest CtsBluetoothTestCases:BluetoothAdapterTests
Change-Id: I22aa5ed81298aa13a243318089cdb5afb6f5e1e2