leaudio: Fix handling HCI event on datapath set
When handling IsoSetupIsoDataPathCb from IsoManager,
state_machine->ProcessHciNotifSetupIsoDataPath with cis handle is called.
That cis_handle should be used to find the ASE which is having
data_path_state in CONFIGURING state.
If not, it is possible that stack incorrectly choose ASE in the same
state but for different cis_handle which will end up in puting state
machine in a wrong state.

Issue can happen very easly when multiple sink and source ASEs are
configured on single device

Bug: 430276201
Flag: EXEMPT, day-to-day fix covered with unit test
Test: atest bluetooth_le_audio_test
Test: atest pts-bot:GMAP/UGG/LLU/BV-44-C -v

Change-Id: I64cda12f0500c2a9e0483f0a105b3012243ab8c5