leaudio: Fix bad audio during GAME sonifications
Some of the applications like Chess.com or Duolingo produce sounds with
 contentType=AUDIO_CONTENT_TYPE_SONIFICATION  usage=AUDIO_USAGE_GAME

This  triggers reconfiguration to GAME context which breaks background
audio if it is in place.

We expect that GAME context will be used for GAMEs which keeps a GAME
track while appliation is opened.

In the future, when Isochronous Parameters Update will be supported, we
might consider changing CIS configuration for such a case if it will be
done seemless. Without IPU, it makes sense to treat GAME SONIFICATIONS
as regular SONIFICATIONS which does not force reconfiguration to GAME.

This improves user experiance.

Bug: 421830251
Bug: 436171822
Test: atest bluetooth_le_audio_client_test
Flag: com.android.bluetooth.flags.leaudio_use_game_sonification_as_regular_sonification
Change-Id: Ic85630e05f87d9bc3e7c9791e53abbecc294a970