leaudio: Fix audio quality on reconnect with Audio HAL issue
This is yet another improvement for the devices which uses Audio HAL
which is not capable to clear metadata.

Issue is that after reconnection there might be short periot of time
when Bluetooth is fooled that Decoding session is about to be used.

With this CL, there is a way to ask audio context type manager to not
take into account given directions. With this API, client.cc can make
use of it and depends on Resume direction and knowledge on the Audio HAL
behavior can omit Decoding direction when needed.

Bug: 431675666
Flag: EXEMPT, day to day fix
Test: atest bluetooth_le_audio_client_test
Change-Id: I9d154788389f8d2922bb34f194f41800d248fe95