Use AttributionSource as source of truth
(Pun not intended)

The uid is coming from `sendingUid` of Messenger and is the same as
AttributionSource.uid. So this CL is no-op.
However, using only one uid for most of the check allows to simplify
permission logic.

The sendingUid is kept for API that doesn't take an AttributionSource,
or for simple alternate permission enforcement (BLUETOOTH_PRIVILEGED)

Bug: 321804999
Flag: com.android.bluetooth.flags.system_server_messenger
Test: atest ServiceBluetoothRoboTests
Change-Id: Ia362049336b969dd08eb951fa3c22766966f8b40