Do not cancel timeout if L2CAP is connecting.
An lcb is allocated when initiating an L2CAP connection
and should be released upon disconnection or connection failure.

However, canceling the connection while it's in progress
happens silently, and the lcb is not released.

This change fixes the issue by deliberately invoking
the timeout alarm to clean up the lcb used for the connection.

Flag: com.android.bluetooth.flags.l2c_not_cancel_timeout
Bug: 436736160
Bug: 429922295
Test: atest bluetooth_test_gd_unit
Change-Id: I86626cbb8ccb05d370cfa6cfece3f303b5717a4f