leaudio: Fix Sink/Source monitor behavior during reconfiguration
So far Sink/Source monitor notification was skip during the
reconfiguration with an assumption that stream will be Resumed after
that. This is wrong assumption as it might never happen and monitor user
would not be notified about stream suspended.

This CL fixes that.

Bug: 432138731
Bug: 440506873
Flag: com.android.bluetooth.flags.leaudio_improve_unicast_monitor
Test: atest bluetooth_le_audio_client_test
Change-Id: Ica12d5ca74de3e28e249d8fdcb4364c438ee24d5