BluetoothSocket: fix finalize issues
Because BluetoothSocket implements finalize, even a non fully
constructed object can have internal method be called later (or never).
Throwing an exception from a constructor should be sufficient to prevent
an object from coming into existence; in the presence of finalizers, it
is not.

About finalize:
  Tip: Don't do it. If you absolutely must, first read and understand
  Effective Java Item 8, "Avoid finalizers and cleaners" very carefully,
  and then don't do it.

Bug: 436815186
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT refactor for test only
Change-Id: Ib93bf28879ac8165f002c018500d9c2f21b27ebf