Remove remove_device_with_connection_manager logic
In current implementation, connection_manager will remove all the background connect clients including the ones which are interested in Targeted Announcements. This will later allow other applications to use Accept List in the controller even it should not be a case as it is LeAudio device and TA shall be used.

Reverting the flag logic here in ag/32536803.

Test: atest bluetooth_test_gd_unit
Bug: 429528446
Bug: 438985552
Flag: com.android.bluetooth.flags.disconnect_acl_on_gatt_timeout
Change-Id: Idb0b2bfce0e1b48c0cc0d3fbe16b102b29ba7b87