Use generic accessor to call adapterService
Instead of rewriting the logic over and over, we can simply write the
wrapper once and call it when needed.
This enforce that the lock is taken.
This enforce consistency.
This enforce return value for a bluetooth off is handle the same way as
a remote exception (caller can not anticipate when bluetooth will be
turned off and behavior should be the same in both cases)

Bug: 311772251
Test: m .
Flag: EXEMPT refactor
Change-Id: I39cb5590a9bd67db597c09118c5b1dce3d2eff00