Optimize map iteration with entrySet in HeadsetClientService
Update getConnectedDevices and getDevicesMatchingConnectionStates to
iterate over the map's entrySet instead of its keySet.

This is more efficient as it avoids the repeated Map.get() lookup
for each key within the loop.

An additional gain is a subtle performance bug in `getDevicesMatchingConnectionStates`.
The original code was calling `mStateMachineMap.get(bd)` inside the inner loop,
meaning the map lookup was being performed multiple times unnecessarily
for the same BluetoothDevice. The new code gets the state machine (sm)
once before the inner loop, which is better.

Bug: 394958283
Test: atest HeadsetClientServiceTest
Flag: EXEMPT refactor no-op
Change-Id: Ib5070337b809901211986d7054ebe9dead2d8ef0