Reset the media connected state when turning off
When Bluetooth is turning off, there is by definition no more media
connected. But the internal variable is not always updated.
We encounter some case where the LeAudio state change wasn't broadcasted
to internal listener and leaving the status desynced.

This can be an issue when the Bluetooth does a ON -> BLE_ON -> ON
transition without going full OFF in between.

Bug: 431184452
Flag: EXEMPT trivial fix
Test: m .
Change-Id: I3ee8c9ee4d7b6ab972977001011710515e522bb9