leaudio: Fix configuration check without codec extentions
With this CL, whenever audio configuration is checked in the LeAudio
group, code is taking group size into acount and not only connected
devices.

Later on, when actually configuration is happening, only connected
devices will be configured anyway.

This is improvement which was introduced with codec extensions but was
not adapted to previous configuration path.

Bug: 441436890
Bug: 441436893
Flag: com.android.bluetooth.flags.leaudio_always_use_group_size_to_check_audio_config
Test: atest bluetooth_le_audio_test
Change-Id: Ifcc4278af5faafd17a663784a79b5f31942e16f3