Unify the timeout in the stack
Bluetooth timeout are a mess. Despite CTS and VTS requesting a 2s start
for the HAL and a 4s start for the Bluetooth entirely. Over the year,
some in between timeout got added to allow a free value to be set. A
single step of the startup can now be defined as more than 8s, without
the enclosing timeout being changed.

This CL will provide a way to increase the CTS timeout from 4s to 8s,
with the counterpart that it will no longer be possible to have a free
value of the timeout.

AKA the following property will no longer be respected:
* ro.bluetooth.ble_start_timeout_delay
* ro.bluetooth.ble_stop_timeout_delay
* bluetooth.gd.start_timeout
* bluetooth.gd.stop_timeout

If a device is abnormally slow, it can now define the following
property: `bluetooth.hardware.degraded_performance_mode.enabled`.

Note: Using `ro.hw_timeout_multiplier` allow extra time without the need
for the new property.

In order to keep retrocompatibility to devices, the property are still
taken into consideration up to (and included) Baklava. Next release will
HAVE to either set the degraded performance mode, or to fix their hal
implementation. The VTS still enforce 2s of HAL startup, and there is no
plan to change it

Bug: 391248793
Bug: 441576259
Flag: com.android.bluetooth.flags.unify_timeout_property
Test: atest CtsBluetoothTestCases
Change-Id: I5dce4be5669b440c7e3113cb507d39cf645fc721