Make obvious that BluetoothGatt service is nonNull
By construction, the service cannot be null, therefor we don't need to
check for nullness.
Inline connect within the constructor as much as possible.

The only reason why registerApp is not inlined into the constructor is
due to lint permission that doesn't parse on the constructor.

Bug: 311772251
Test: m .
Flag: EXEMPT PURE_REFACTOR
Change-Id: I22b067fd56e26a87efdc6d2e7d5dfc40b6de9c14