snoop_logger: Fix the logs directory access
Fix a weird case probably only happens during tests; the old behavior is
crashing so this doesn't introduce risks.

When logged in as a non-primary user, the euid of bluetooth service
would not be AID_BLUETOOTH and it relies on the *group* permission to be
rwx for AID_BLUETOOTH to add/remove the files inside.

This patch fixes this by setting the correct umask (was 0077) and
changing the group to AID_BLUETOOTH. More properly we should set the
user to AID_BLUETOOTH as well but the process doesn't have the
permission doing that.

Bug: 443679212
Test: m com.google.android.bt
Test: floss/build/build-in-container.py
Flag: EXEMPT BUGFIX
Change-Id: I0636c67aeb050045e54e435188aba3ae3682f917