Updated app lint-baseline with manual suppression of @RequiresPermission
This is a manually curated list of all lint errors from the new
`RequiresPermissionDetector` that overlap / is mapped to a
`@SuppressLint("AndroidFrameworkRequiresPermission")`. All existing
suppressions are being baselined which will allow the enabling of the
new lint tool.

Exact details of all errors with class names and lines can be found in:
https://buganizer.corp.google.com/issues/409768276#comment28

Bug: 409768276
Test: atest --host BluetoothLintCheckerTest:com.android.bluetooth.lint.test.RequiresPermissionDetectorTest
Flag: NONE LINT_ONLY
Change-Id: I8bf1d4696b467e7c317895b7a0d4a142190cdeb5