Force changed event for new state
If the playback state has changed between the last changed event and
the interim, we should send a new changed event as some devices don't
treat interim as a change of state. This is done by sending the last
state as interim, and the new state as changed.

Bug: 441441368
Flag: EXEMPT BUGFIX
Test: m .
Change-Id: I393cc63e220a5e94a828f9bfba8fb46e1b0bbbfa