Handle null and empty Voice Assistant names
If a Voice Assistant app is not selected as a default app
in the Settings, the Voice Assistant name will be null or
an empty string.

In this situation, the VA name characteristic value will
be set to "None" to comply with the specification requiring
a non-empty value for the Voice Assistant name.

Consequently, the VA Session State will be set to VA SESSION UNAVAILABLE.

Test: mmm packages/modules/Bluetooth
Test: VAPS manual and IOP test
Flag: EXEMPT BUGFIX
Bug: 445114212

Change-Id: I73cb52c761987a0baab6ec06f7ad96f4dc3226e1