RequiresPermissionDetector: Support ...MultiplePermissions broadcasts
The RequiresPermissionDetector lint check did not recognize the
sendBroadcastMultiplePermissions or
sendBroadcastAsUserMultiplePermissions methods. This meant it couldn't
validate the permissions parameter for these new API overloads.

This allows the lint to correctly compare the broadcast's action
permission with the permissions provided in the method call,
regardless of which overload is used.

Bug: 409768276
Test: atest --host BluetoothLintCheckerTest:com.android.bluetooth.lint.test.RequiresPermissionDetectorTest
Flag: NONE LINT_ONLY
Change-Id: Ib20596ca2cb64bb39d206960ab22073796e6915e