Send VA Session State notification for all the devices in a group
Currently, VA Session State notifications are only sent to the first
device in a group.
This limitation occurs because notifications are triggered solely by
a change in the VA session state. Once the notification is sent for
the initial device, the session state is updated, and subsequent
devices in the group do not receive a notification as the state remains unchanged.

This change ensures that all devices within a group receive VA Session
State notifications.

Test: mmm packages/modules/Bluetooth
Test: VAPS manual and IOP test
Flag: com.android.bluetooth.flags.leaudio_vaps_improvements
Bug: 444505646
Bug: 444029563

Change-Id: I6df8b1dabfcd98579f9db1982f2f8b6fa86be9a7