RequiresPermissionDetector: Check super methods for permission annotations
The PermissionEnforcementVisitor was only checking the currently called
method for `@RequiresPermission` or `@EnforcePermission`. This caused
it to miss required permissions if the annotation was defined on a
super method (e.g., in an AIDL interface).

Real example: In `HidDeviceServiceBinder` and `HidHostServiceBinder`,
2 functions that have a `@RequiresPermission(BLUETOOTH_CONNECT)` were
calling another public function that had the same requirement for code reuse.

Bug: 409768276
Test: atest --host BluetoothLintCheckerTest:com.android.bluetooth.lint.test.RequiresPermissionDetectorTest
Flag: NONE LINT_ONLY
Change-Id: Ifef9f3de7421c2dcb3d11aa190d8efadfb1009b8