Fix RequiresPermissionDetector to scan inside lambdas
When a `sendBroadcast` call, it would look for the `Intent` action
permission but with fail to find it if it was inside a nested forEach
lambda. This traverses all method arguments in lambdas.

A real example:
Within RemoteDevices.java#void reset(), each call is nested inside a
`mDevices.forEach(` loop with a lambda. These were failing to be caught
but now are also detected.

Bug: 409768276
Test: atest --host BluetoothLintCheckerTest:com.android.bluetooth.lint.test.RequiresPermissionDetectorTest
Flag: NONE LINT_ONLY
Change-Id: I776ca8b6cb6da34f53e9213b147c89d2e2378934