[Samsung Convergence] Only call ScanManager#stopBatchScan if it is a batch scan client
In AOSP, when `stopScan` is called, the `stopBatchScan` logic is
executed even if it is not a RegularScan. This results in
`stopBatchScan` being performed for suspended scans as well. To reduce
unnecessary operations, modify the system to execute `stopBatchScan`
only when there is a `batchScanClient`.

Bug: 446532441
Bug: 445817393
Test: atest ScanManagerTest
Flag: com.android.bluetooth.flags.stop_batch_scan_only_if_batch_client
Change-Id: If40d58f642b4e53cec6f3d8afdd12a76338f409f