Build bluetooth tests with static libc++ and version script
The test gets compiled against the new libc++.so but the test runs on
older platform versions. Mainline pushes only the test so the test loads
the old libc++.so on the older platform. This causes undefined symbol
issues because the old libc++.so will not have new symbols. The fix is
compiling the test statically and the easiest way was to add it to
mts_defaults since it is used by the 30+ tests. You need to add the
version script to fix asan issues.

Bug: 445941769
Flag: EXEMPT BUGFIX
Test: m && ABTD Run
Change-Id: Ib25122939bfd7d25621757636987090ccef0f793