ActiveLogs: Format dump output as an aligned table
For a sample output, see:
https://buganizer.corp.google.com/issues/446315696#comment2

The output of the `ActiveLogs.dump()` function was misaligned and
difficult to read due to variable-length package names. This made
parsing the logs unnecessarily difficult.

This change refactors the `dump()` function to print a clean,
symmetrical table instead of just printing the `toString()` of
each log.

- Calculates the `packageColWidth` dynamically based on the longest
  package name.
- Uses `padEnd()` to left-align and pad the TIMESTAMP, PACKAGE, and
  ACTION columns.
- Adds a table header and a separator line for improved readability.
- Changes `ActiveLog` properties from `private` to `internal` to allow
  the `dump()` function to access them directly for formatting.

Bug: 446315696
Test: atest ServiceBluetoothRoboTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: Iebd3d7a323720b9942ee2b2560c8a223f74a6fd6