Better encapsulation of ConnectableProfile
AvrcpController never had any entry in the database to save a persistent
connection policy.
And the associated method are never called.
Cleaning dead code and making this explicit

Other profiles are connectable but doesn't expose the 'connect' api.
Since the whole storage exist and there are possibility to save the
persistent state, keeping them as ConnectableProfile... for now

Flag: EXEMPT PURE_REFACTOR
Test: m .
Bug: 430166215
Change-Id: I6a6a696419a2f51899ce2631044e3fd79b3cddb4