Replace base::Callback and base::Closure with base::Once* classes
base::Callback and base::Closure are deprecated and should be replaced.
Replace them in LeAudioClient::Initialize and related files.
According to the comment in bta_gattc_api.cc `initCb` is expected to
actually be a “one time callback” (and handles it as such) and
`hal_2_1_verifier` is only called once.

Bug: 303363630
Test: atest packages/modules/Bluetooth
Flag: EXEMPT PURE_REFACTOR

Change-Id: I74dbf0fd4ea338b7eae864e55f56d271a8769d46