BASS Client: store Sink Metadata when receiver sink syncs to BISes
So far, we stored which channels are selected only when we received
this information from settings - when the broadcast is resumed after
pause, so we can send Modify Source with correct BISes to sync to.

However, this can lead to a failure in handover scenarios, when
Broadcast wasn't paused and resumed before. In that case, no channel
would be selected, and when we exit Unicast back to Broadcast, no
channel selected would mean that we have nothing to resume.

This affects path in `resumeReceiversSourceSynchronization`, where
receivers with no channels selected are skipped from resuming.

Test: atest BassClientServiceTest
Bug: 425589971
Bug: 419338246
Flag: com.android.bluetooth.flags.leaudio_broadcast_stop_big_monitoring_based_on_bis_sync
Change-Id: I44dc0b4f653e6641283f0dfa6c6c0701dfe22a3f