leaudio: Fix handling single device with invalid CSIS service
There are devices on the market which are LeAudio headsets with a set
size set to 1 but SIRK set to all 0x00 which is out of spec behavior.

Android at the moment will not connect such LeAudio headset but with
this CL, CSIS is considered as not valid and will not be connected but
LeAudio headset will be usable as a single device.

Bug: 420393034
Bug: 447588233
Test: atest bluetooth_le_audio_client_test bluetooth_csis_test
Flag: com.android.bluetooth.flags.csis_quirk_for_single_device_with_sirk_all_zeros
Change-Id: I2deece9a318da1e371a7498d494dc288422ce3ae