[le audio] Add existing source from peer device to newly connected device
When BASS connected a grouped device, it needs to check if there is an existing source
on the peer device, and then proceeds with adding the peer source to this newly connected device.

With this change, BASS ensures the grouped devices have the same source in private broadcast.

Bug: 447646642
Bug: 443874419
Flag: com.android.bluetooth.flags.leaudio_broadcast_add_source_from_peer_device
Test: atest BassClientServiceTest
Change-Id: I9ef8d51536ee34d7d4c0f76b3ce79aa2467255ba