Coalesce ACL connection state change broadcasts.
Set broadcast delivery group when sending ACTION_ACL_CONNECTED
and ACTION_ACL_DISCONNECTED broadcasts to allow coalescing
multiple connection and disconnection events that are not yet
delivered to the receiver apps.

This sets a delivery group matching key based on the device
address and transport, and sets the policy to deliver only the most
recent event. This should reduce the number of broadcasts that are
delivered.

Bug: 413570777
Bug: 448673677
Test: atest ./android/app/tests/unit/src/com/android/bluetooth/btservice/RemoteDevicesTest.java
Flag: com.android.bluetooth.flags.coalesce_acl_connection_broadcasts
Change-Id: I1dc5aa6d596f00d9d2f9bd9c45e36529b82c7d8b