Allow ScanNativeCallback to unblock the scan thread
This prevents a deadlock happening where the scan thread was put on
hold and the callback operations was being added to the same thread
which would prevent the unblocking of thread via `callbackDone()`.

For #onBatchScanReports, this does change the order of execution when
the scan thread flag is true. Previously, it would interrupt
ScanManager#flushBatchResults with
ScanController#onBatchScanReportsInternal and wait for it to finish
first before continuing. Whereas now flushBatchResults finishes before
onBatchScanReportsInternal is triggered. The biggest difference is
ScanManager#setBatchAlarm is called prior to the delivery of
ScanController#deliverBatchScan which at the moment, I cannot see a
functional difference.

Bug: 397863857
Fix: 448051603
Test: atest com.android.bluetooth.le_scan
Flag: com.android.bluetooth.flags.scan_controller_thread
Change-Id: Ie0c173ef01e0d6c11a09847e24858b8849043ad9