VC: Skip redundant volume/mute writes to avoid pending operations
When the VC stack is notified by a device about a change in volume
control state, it must forward this update to other group members.
However, if devices synchronize by themselves and update their state
before the stack forwards the operation, the stack will not be notified
again. This can leave a pending operation stuck.

This situation is especially likely when the volume is set to 0 and mute
is applied. When a single notification contains both a volume change and
an unmute. The stack then attempts to forward them as two separate
operations. In such cases, the issue may occur even without
device-to-device synchronization, for example when buds autonomously
unmute on volume increase.

This change fixes the problem by checking if the current value matches
the queued operation. If so, the operation is skipped.

Bug: 443437199
Bug: 448577346
Flag: com.android.bluetooth.flags.vcp_skip_redundant_operation_writes
Test: atest --host bluetooth_vc_test
Change-Id: I7758037a85375c1f9379a4e5501937e3594139b5