Convert ScannerApp to a top-level Kotlin class
This is in preparation for a more detailed scanner app history.
Currently, we only output the scan data of the active scan and scan
mode of only the latest scan. We are planning a more detailed approach
that will allow tracking more details for app scanning abuse analysis,
battery and performance impact.

Also see: b/327293682

All fields are added as `@JvmField` to prevent a massive change across
all files due to new variable naming by Kotlin when accessing from Java

Ideally, `ScannerApp` should be a data class and so the next step is to
look into that possibility.

Bug: 448760632
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I9cf92c751e9ed54c60939030078975ee6f60b40c