Remove all `@JvmField` annotations and finalized fields in ScannerApp
This also allows `callback`, `info`, and `appScanStats` to be final
due to the tests using the pattern `doReturn(foo).when(app).getFoo()`.

All variable names have been modified to drop the Java specific `m`
prefix as well. When called from Java, this requires to replace direct
field access with property accessors such as `getId()`.

Also prefer `app` in a few place instead of redundant `scannerApp` to
make it consistent with the rest of the class.

Bug: 448760632
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT PURE_REFACTOR
Change-Id: I5a556648cc23bbb8715d721365f075dc88097fd8