Fix ScanControllerTest slowness when scan_controller_thread is on
`forceRunSyncOnScanThread` creates a deadlock when using the TestLooper
in conjunction with `CompletableFuture`. The test thread is stuck
waiting inside `future.get()`. The Runnable that would complete the
future can only run if the TestLooper is advanced. But since the test
thread is blocked, calling even `mLooper.dispatchAll()` does not
unblock it because of the `CompletableFuture` lock.

Bug: 397863857
Test: atest ScanControllerTest
Flag: com.android.bluetooth.flags.scan_controller_thread
Change-Id: I3e4e7e2601a07e2dc3259897dfccb63d4473e3fa