Re-add device address to audio state changed events from native
Problem: We want to make individual devices manage their own audio state
instead of trying to centrally own it. To do this, we need context of
which device an audio state event is coming on behalf of. We used to
have this, but it was removed in a refactor of StackEvent as dead code.

Solution: Add it back on. Send the event to the state machine to log.

Flag: EXEMPT PURE_REFACTOR
Bug: 450371968
Test: atest com.android.bluetooth.a2dp_sink
Change-Id: I868b3e80b4e3c78bd067c95ca50cb6c577c3536c