SuspendManager: Allow some advertisement parameters to be null
Some of the advertising parameters can be nulled, but previously they
are not declared as such in the suspend manager. This causes null
pointer exceptions when a null object is passed from the Java code.

Set some of the nullable parameters as nullable.

Bug: 449265206
Bug: 433136730
Test: m -j
Test: manually suspend and resume
Flag: com.android.bluetooth.flags.adapter_suspend_advertisement
Change-Id: Ic1f2265d44dbb5793d16547aba1563542b04a986