Fix user switch handling in BluetoothManagerService
BluetoothManagerService does not clears the mNextUser when
user switch happens from user 0 to user 10 on reboot. This
leads to below issue:

> After disabling bluetooth and rebooting, bluetooth fails to
turn on with error 'user switch in progress'

Ensure mNextUser is cleared when current user equals
mNextUser in the BluetoothManagerService.

There is 2 situation in which we may re-switch on the same user, both
can lead to a 'lock' system. If Bluetooth is OFF and
* we switch to user B then back to user A before performing the switch
* We never switch to the current user (HSUM device) and we received the
  first switch notification before handling a bluetooth start (or maybe
  the start wasn't enabled)

Bug: 449926982
Fix: 449926982
Flag: EXEMPT BUGFIX
Test: atest ServiceBluetoothTests
Test: manual: After reboot able to turn on the Bluetooth

Change-Id: I0efa7350991c1fd4623947efed69c025fb759eff
Signed-off-by: Panda, Bharat B <bharat.b.panda@intel.corp-partner.google.com>
Signed-off-by: Bhadouria, Aman <aman.bhadouria@intel.corp-partner.google.com>
(cherry picked from commit bf19b237a57db5c568247c323d16a88272c60a88)