SetName: Can be set when Bluetooth if OFF
Allow consistent behaviour of the 'setName' API.

This API allow to set the local name of the device as it will be seen by
remote Bluetooth devices.
There is only one API exposed, but this name can be set by a real user
in 2 different location. In the Bluetooth settings, that require
Bluetooth to be ON to be accessed, and when changing the whole device
name.
This second location can only change the Bluetooth name when Bluetooth
is ON, without having a way to enforce it. As a result, if you only
change the name here, you end up with inconsistent behaviour where the
name is sometimes set and sometimes not.

By moving the logic of setting the name in the system server (which was
already having some logic to save the name on disk), we now allow the
name to be set anytime. If Bluetooth is OFF, it will be used the next
time the Bluetooth is started.

Also, we are adding a way to reset the name to it's default value, by
either calling the API with null or with an empty string, we will
default to what was the device name at build time which is by order of
priority:
* bluetooth.device.default_name system property
* Setting.Global.DEVICE_NAME
* ro.product.model
* "Android"

FYI: Bluetooth name is UTF8 encoded. It's size in bytes must be smaller
than 248 bytes

Bug: 448709503
Bug: 437058714
Bug: 342634832
Bug: 279180695
Flag: com.android.bluetooth.flags.set_name_in_system_server
Test: atest BluetoothJavaUnitTests
Test: atest CtsBluetoothTestsCases
Change-Id: I6a6a69645263672c9e1f9c6c3f16dd19269f0f29